thermocycler-gen2: add error control (#534)
* refactor: pull some repetitive includes out of tests

* refactor: improve tests

Take the same approach to the tests as the heater-shaker, condensing
host comms message checking.

* feat: add error handling gcodes to tcgen2

They don't do anything right now, but we can copy the gcode parsing and
response parts from the heater-shaker because they'll work the same way.
The test structure is a bit different though.

* feat: thermal plate get-error

* remove from motor

* feat: add error handling to thermal plate

The error handling works like this:

- We will save an "interrupted-by" error that persists until the next
GetErrorState message, even if the error that caused it originally goes
away. This is important so that we don't miss transient errors that
cancelled any ongoing work
- We get an ongoing error state in preference to an interrupted-by
- When we clear the error state, it's not really committed (i.e., the
LEDs don't change) until the next temperature conversion, to avoid any
weird flickering. Not that you'd see that in practice for now, since
it's a test command.
- When we set the manual error, it may get overridden by a subsequent
real error.

* Format lots of stuff

* lid: formats and some std::ignore twiddling

* feat: implement error management in lid

* fix: allow messages on subsequent HCT runs

If the host comms task sends responses too fast it'll clobber the
previous buffer. Avoid this with a nice little sleep since the stm32 usb
class doesn't appear to let you see when a buffer has been sent.

* fix some state passing and led setting

* fix lint

The null pointer thing might be this bug:
https://bugs.llvm.org/show_bug.cgi?id=46235 the others are just a little
gross. I might want to disable the constant array index one because if
you're not going to tolerate exceptions, as we don't, then it's not a
good idea.

* this test was broken oops

* format said test