add usb transmission backpressure (#536)
* heater-shaker: add a wait after send

This looks gross and is gross but fixes an issue where if the host comms
task has two messages to send in quick succession, you can end up
running again swapping out the tx buffer before it actually hits the wire.

* add usb tx backpressure

We can actually get whether the usb is busy out of the usb cdc
machinery. We can use that to handle delaying and retrying if the usb is
busy, and waiting until the usb transaction completes to actually send data.

* limit retries

* format