fix(thermocycler-gen2): reset lid pos cache if err (#538)
Our fix to EXEC-1451 was to detect if the lid status switches detected
that we weren't actually closed after the seal extended and return an
error. This works great. The problem is that because this check happens
in the lid state machine, the lid _hinge_ state machine has already done
its work and cached that the lid is closed. Then, when the robot retries
the command, the lid switches report in-between but we fall back to the
cached value... which is "closed"... so we immediately and incorrectly
succeed.

By clearing the position cache in the path that raises the error, we
make sure that a subsequent close command will actually run (including
raising the same error again if the error happens again) and thus the
robot's retry and eventual error mechanisms will take us home.

Closes EXEC-1451